CMake: greentea: Port mbed-drivers-ticker to CTest
Call add_test in mbed_greentea_add_test and specify mbedhtrun as the
command to run. To allow mbedhtrun to communicate with the DUT, the user
is required to pass in -DMBED_TARGET_SERIAL_PORT, and
-DMBED_TARGET_MOUNT_POINT so mbedhtrun can find the device.

Also add an MBED_HTRUN_OPTIONAL_ARGUMENTS variable which takes a
semicolon separated list of optional arguments to forward to htrun.

Temporary checks have been added to mbed_greentea_add_test to keep the
old flow working until we port all of the greentea tests to CTest.