CMake: tests: Support skipping unsupported test with reason
Add a new argument `TEST_SKIPPED` to `mbed_greentea_add_test()` to
indicate a test is skipped and give a reason (e.g. the test does not
support the target and/or configuration used).

This is achieved with a custom target in place of an executable
while keeping the same target name. The skipping message is printed
at both configuration and build stages.